% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escapstructuralgravity-package.R
\docType{package}
\name{escapstructuralgravity-package}
\alias{escapstructuralgravity}
\alias{escapstructuralgravity-package}
\title{escapstructuralgravity: Open Trade Statistics Model App}
\description{
Wrappers around gravity models for interactive model building. This is meant as a production grade Shiny app and not as end user app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tradestatistics/visualization-with-shiny}
  \item Report bugs at \url{https://github.com/tradestatistics/visualization-with-shiny/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mavargas11@uc.cl}

}
\keyword{internal}
